Proxy works now
If the proxy randomly stops working in the future, play around with using <hosturl> and '/oauth-api' for each of the request URL, the context URL, and in `pathRewrite`.

At the time of this commit, <hosturl> is https://migrationtooldev.com. The request URL is the URL used in the HTTP request. The context URL is the value of `context` in proxy.conf.ts, and `pathRewrite` also from proxy.conf.ts.